[feat] splash 페이지 구현
